feat: add renderWithRouterAndRedux utility and update tests to use Redux
