+++ src/pages/searchPage/components/apiResults/ApiResults.test.tsx
@@ -2,11 +2,11 @@ import { describe, it, expect } from "vitest";
 import { act, screen } from "@testing-library/react";
 import ApiResults from "./ApiResults.tsx";
 import type { Book, BookSearchResponse } from "@api/book/models.ts";
-import renderWithRouter from "@/test-utils/renderWithRouter.tsx";
+import renderWithRouterAndRedux from "@/test-utils/renderWithRouterAndRedux.tsx";
 import LocationDisplay from "@/test-utils/LocationDisplay.tsx";
 import { QUERY_DETAILS_ID } from "@pages/detailsPage/DetailsPage.constants.ts";
 
-const render = renderWithRouter;
+const render = renderWithRouterAndRedux;
 
 describe("ApiResults", () => {
   it("shows a spinner when result is null", () => {
